+++ src/com/paa/algoritmos/MergeSort.java
@@ -7,13 +7,13 @@ package com.paa.algoritmos;
 
 /**
  * @author jaovctr
- * @version 1.0
+ * @version 1.1
  * @since 1.0
  * Algoritmo Merge Sort
  * TODO comparações
  */
 public class MergeSort {
-	private static int comparacoes=0;
+	private static long comparacoes=0;
 	
 	/**
 	 * @return the comparacoes
